Autism_care_hub-website/Autism_care_hub/backend/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.36.0
requests==2.26.0
